backend/requirements.txt
asgiref==3.5.0
Django==4.0.4
djangorestframework==3.13.1
psycopg2==2.9.3
python-environ==0.4.54
pytz==2022.1
sqlparse==0.4.2
tzdata==2022.1
